vscode: Install extensions during bootstrap
I can generate an installed extension list with

    code-oss --list-extensions | sort > ./configuration/vscode.extensions

This allows me to also restore the extensions during bootstrap.

Unfortunately, the `--install-extension` does not take multiple
arguments and is quite slow and verbose on its own.

I played around with the idea of using

    command="$(awk 'BEGIN { printf "code-oss" } { printf "--install-extension " $1 } END { printf "\n" }' <./configuration/vscode.extensions)"

to generate the command for quicker execution but it was simpler and
more understandable to simply iterate through the list.